ipgeo/requirements.txt
###### Requirements without Version Specifiers ######
requests
pwntools
colorama
ipaddress